+++ scripts/mods/BetterInventory/BetterInventory_data.lua
@@ -1,4 +1,4 @@
-local MOD_VERSION = "3.3.1"
+local MOD_VERSION = "3.4.0"
 local mod = get_mod("BetterInventory")
 local DEFAULT_OPERATIVE_SLOT_CAPACITY = 10
 local MAX_REASONABLE_OPERATIVE_SLOT_CAPACITY = 64
@@ -590,6 +590,12 @@ return {
 						default_value = false,
 					},
 					{
+						setting_id = "enable_inventory_search_melk",
+						tooltip = "enable_inventory_search_melk_tooltip",
+						type = "checkbox",
+						default_value = true,
+					},
+					{
 						setting_id = "inventory_search_non_match_behavior",
 						tooltip = "inventory_search_non_match_behavior_tooltip",
 						type = "dropdown",
@@ -663,6 +669,34 @@ return {
 								tooltip = "inventory_search_spacing_reopen_tooltip",
 								type = "numeric",
 								default_value = 50,
+								range = {0, 96},
+							},
+							{
+								setting_id = "inventory_search_melk_limited_top_padding",
+								tooltip = "inventory_search_spacing_reopen_tooltip",
+								type = "numeric",
+								default_value = -40,
+								range = {-50, 64},
+							},
+							{
+								setting_id = "inventory_search_melk_limited_bottom_padding",
+								tooltip = "inventory_search_spacing_reopen_tooltip",
+								type = "numeric",
+								default_value = 40,
+								range = {0, 96},
+							},
+							{
+								setting_id = "inventory_search_melk_multi_top_padding",
+								tooltip = "inventory_search_spacing_reopen_tooltip",
+								type = "numeric",
+								default_value = -40,
+								range = {-50, 64},
+							},
+							{
+								setting_id = "inventory_search_melk_multi_bottom_padding",
+								tooltip = "inventory_search_spacing_reopen_tooltip",
+								type = "numeric",
+								default_value = 45,
 								range = {0, 96},
 							},
 						},
@@ -1381,6 +1415,18 @@ return {
 						setting_id = "melk_views_group",
 						type = "group",
 						sub_widgets = {
+							{
+								setting_id = "enable_melk_limited_grid",
+								tooltip = "enable_melk_limited_grid_tooltip",
+								type = "checkbox",
+								default_value = true,
+							},
+							{
+								setting_id = "enable_melk_multi_operative_grid",
+								tooltip = "enable_melk_multi_operative_grid_tooltip",
+								type = "checkbox",
+								default_value = true,
+							},
 							{
 								setting_id = "melk_auto_favorite_purchased_items",
 								tooltip = "melk_auto_favorite_purchased_items_tooltip",